Drlll Description componentized for description screen and description modal and deleted image carousel
